+++ script/patch-custom.sh
@@ -28,3 +28,14 @@ sed -i '/\/etc\/init\.d\/tailscale/d;/\/etc\/config\/tailscale/d;' $OPENWRTROOT/
 
 # Fix docker bridge network
 sed -i 's/list blocked_interfaces '\''wan'\''/#&/' $OPENWRTROOT/feeds/packages/utils/dockerd/files/etc/config/dockerd
+
+# luci-app-dockerman: hide config buttons in overview page
+if [ "$BRANCH" == "v23.05.2" ]; then
+  patch_url="https://github.com/openwrt/luci/commit/568b44edb3bf545e424e7c1a3f162e753b190113.patch"
+  
+  pushd $OPENWRTROOT/feeds/luci
+    wget $patch_url
+    git apply $(basename $patch_url)
+    rm $(basename $patch_url)
+  popd
+fi